Remove 0/8 and 240/4 from bad_ipv4 ranges
These address ranges are reserved by RFC and expected to be blocked by
most routers today (though not necessarily by most hosts, as some
implementers have removed these special cases).  People, including but
not limited to the IPv4 Unicast Extensions Project, are interested in
testing their actual behavior on the Internet in the future.  This
patch removes these ranges from the bad_ipv4 function so that it is
possible to ask a probe to perform a test related to addresses in these
ranges.